+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.6
 
 - Nüshu Light: Use a lighter green (same as GitHub Light Default)
+- Nüshu Light/Dark: Use a custom color with borders (17 instances in each theme)
 
 ## 2.1.5